+++ unskip-and-fix.sh
